Got closer to implementing Subscription
Added Observable for Podcast and Subscription identifier, since an
individual Podcast doesn't need a whole Subscription. Just the
identifier itself to denote state of subscription.

Still need to implement the PodcastService subscription and
unsubscription. Since this is a modification from the application
itself, the database modifications need to happen on a background
thread, so we need to wrap the calls in a Single