Tried to alter autoincrement, but failed.
Tried to make tests more robust with pseudorandom autoincrement values
for each table so that different tables would be less likely to have the
same identifiers, but this broke a lot of other things for some reason I
don't know. It's possible that I have bad code, or it's possible that
messing with SQLite internals was bad. I'm going to guess my code is
fine. We'll keep this branch around in case I'm wrong. :)