Separated local vs remote PodcastService ops
We need to separate local vs remote PodcastService operations so that UI
tests can know what kind of results we're waiting for. Some tests are
waiting just for local results of any kind, but other tests will wait
for remote results. Similarly, we'll want to show a spinner if there are
no local results, but the remote operation is pending.

However, order matters here. If we keep these operations separate, we
lose transactional power. Instead, we should make a union type for local
vs remote updates.