Need to refactor the tables with vector clocks
Each table needs a notion of its monotonically increasing version so
that a given observer can know that its affects have been applied. This
means that every table in the database will need a row in a vectors
table, and each of those rows will need to increase with every
transaction. We can reset these to zero when the application starts.