Updated dependencies, fixed some races
ConstraintLayout is the main culprit. A Group doesn't synchronously mark
its referenced views as visible when it becomes visible. As easy
workaround is to not use Groups that way, but there must be a way to
hook into Android layout logic to know that the layout has stabilized
before marking the system as Idle.