Folded facade mutations into transaction factory
We want the actions we perform on the transaction before we send it to
the renderer to be obvious, so rather than using a generic transaction
with lots of individual operations (which we might forget to call) we
create purpose-built interfaces with a single atomic method accepting
the state and returning a `Completable`.

Did this on `RxValueFragment`. Still need to do it to
`RxAsyncValueFragment`.