Finished testing episode versions.
We can do per-row versions on episodes since they only have a to-one
relationship. Any to-many relationships will need versions on join
tables.

Per-row versions will work on episodes by updating versions of all
episodes for a given podcast when we CRUD any one episode for that
podcast.

We'll get rid of the MetaTable.